+++ shaders/lib/lighting/simple_fsh_lighting.glsl
@@ -19,6 +19,29 @@ float getShadowBrightness(vec3 viewPos, vec3 normal, float lightDot, float ambie
 
 void doSimpleFshLighting(inout vec3 color, float blockBrightness, float ambientBrightness, float specularness, vec3 viewPos, vec3 normal) {
 	
+	#if AMBIENT_CEL_AMOUNT != 0
+		ambientBrightness = sqrt(ambientBrightness);
+		ambientBrightness = mix(ambientBrightness, floor(ambientBrightness * 3.0 + 0.5) / 3.0, AMBIENT_CEL_AMOUNT / 100.0);
+		ambientBrightness *= ambientBrightness;
+	#endif
+	#if BLOCKLIGHT_CEL_AMOUNT != 0
+		blockBrightness = sqrt(blockBrightness);
+		blockBrightness = mix(blockBrightness, floor(blockBrightness * 3.0 + 0.5) / 3.0, BLOCKLIGHT_CEL_AMOUNT / 100.0);
+		blockBrightness *= blockBrightness;
+	#endif
+	
+	#if defined OVERWORLD || defined END
+		float lightDot = dot(normalize(shadowLightPosition), normal);
+		#ifdef SHADOWS_ENABLED
+			float lightDotLift = 0.25;
+		#else
+			float lightDotLift = 1.0;
+		#endif
+		lightDot = lightDotLift * 0.5 + (1.0 - lightDotLift * 0.5) * lightDot;
+	#else
+		float lightDot = 1.0;
+	#endif
+	
 	// night saturation decrease
 	#ifdef OVERWORLD
 		float nightPercent = 1.0 - dayPercent;
@@ -29,17 +52,28 @@ void doSimpleFshLighting(inout vec3 color, float blockBrightness, float ambientB
 		color += nightPercent * 0.06;
 	#endif
 	
-	#if defined OVERWORLD || defined END
-		float lightDot = dot(normalize(shadowLightPosition), normal);
-	#else
-		float lightDot = 1.0;
-	#endif
-	
 	#ifdef END
 		ambientBrightness = 1.0;
 	#endif
 	
 	vec3 ambientLight = getAmbientLight(ambientBrightness, lightDot);
+	
+	vec3 normalForSS = mat3(gbufferModelViewInverse) * normal;
+	// +-1.0x: -0.5
+	// +-1.0z: -0.15
+	// +1.0y: +0.325
+	// -1.0y: -0.65
+	normalForSS.xz = abs(normalForSS.xz);
+	normalForSS.y *= sign(normalForSS.y) * -0.25 + 0.75; // -1: *1, 1: *0.5
+	float sideShading = dot(normalForSS, vec3(-0.5, 0.65, -0.15));
+	float brightForSS = max(blockBrightness, ambientBrightness);
+	sideShading *= mix(SIDE_SHADING_DARK, SIDE_SHADING_BRIGHT, brightForSS * brightForSS) * 0.75;
+	ambientLight *= 1.0 + sideShading;
+	#ifdef NETHER
+		blockBrightness *= 1.0 + 0.75 * sideShading;
+	#else
+		blockBrightness *= 1.0 + sideShading;
+	#endif
 	
 	#if BLOCK_BRIGHTNESS_CURVE == 2
 		blockBrightness = pow2(blockBrightness);
@@ -53,11 +87,12 @@ void doSimpleFshLighting(inout vec3 color, float blockBrightness, float ambientB
 	
 	float shadowBrightness = getShadowBrightness(viewPos, normal, lightDot, ambientBrightness);
 	shadowBrightness *= min((sunLightBrightness + moonLightBrightness) * 5.0, 1.0);
-	shadowBrightness *= ambientBrightness;
+	shadowBrightness *= ambientBrightness * ambientBrightness;
 	float rainDecrease = rainStrength * dayPercent * (1.0 - WEATHER_BRIGHTNESS_MULT);
 	shadowBrightness *= 1.0 - rainDecrease;
 	
 	vec3 skyLighting = shadowcasterLight * shadowBrightness;
+	skyLighting *= 1.0 + 0.5 * sideShading;
 	ambientLight *= 1.0 - shadowBrightness;
 	
 	vec3 lighting = ambientLight + skyLighting;
@@ -72,14 +107,18 @@ void doSimpleFshLighting(inout vec3 color, float blockBrightness, float ambientB
 		float specular = max(dot(reflectedDir, lightDir), 0.0);
 		specular *= specular;
 		specular *= specular;
-		specular = 1.0 - (1.0 - specular) * (1.0 - specular);
+		specular *= specular;
 		specular *= 1.0 - betterRainStrength;
-		vec3 specularColor = shadowcasterLight * (sunAngle < 0.5 ? vec3(1.0, 1.0, 0.5) : vec3(0.5, 0.7, 0.9) * 0.75);
-		specularness *= 1.0 - getSaturation(color);
-		lighting += specularColor * specular * (0.1 + 0.9 * specularness) * shadowBrightness;
+		vec3 specularColor = shadowcasterLight * (sunAngle < 0.5 ? vec3(1.0, 1.0, 0.6) : vec3(0.5, 0.7, 0.9) * 0.75);
+		#if PBR_TYPE == 0
+			specular *= 1.0 - 0.25 * getSaturation(color);
+		#endif
+		lighting += specularColor * specular * (0.1 + 0.5 * specularness) * min(shadowBrightness * 64.0, 1.0) * min((sunLightBrightness + moonLightBrightness) * 5.0, 1.0);
 	#endif
 	
-	blockBrightness *= 1.2 - min(getLum(lighting), 1.0);
+	blockBrightness *= 1.25 - min(getLum(lighting), 1.0);
+	lighting *= 1.0 - 0.25 * blockBrightness;
+	
 	#ifdef OVERWORLD
 		blockBrightness *= 1.0 + ambientBrightness * moonLightBrightness * (BLOCK_BRIGHTNESS_NIGHT_MULT - 1.0);
 	#endif
@@ -96,6 +135,7 @@ void doSimpleFshLighting(inout vec3 color, float blockBrightness, float ambientB
 	}
 	vec3 nightVisionMin = vec3(betterNightVision);
 	nightVisionMin.rb *= 1.0 - NIGHT_VISION_GREEN_AMOUNT;
+	nightVisionMin *= 1.0 + 0.5 * sideShading;
 	lighting += nightVisionMin * (1.0 - 0.25 * getLum(lighting));
 	
 	#if DO_COLOR_CODED_GBUFFERS == 1