+++ shaders/style_vanilla.glsl
@@ -89,11 +89,11 @@
 
 #if VANILLA_AO_BRIGHT == -1
 #undef VANILLA_AO_BRIGHT
-										#define VANILLA_AO_BRIGHT 0.9
+										#define VANILLA_AO_BRIGHT 1.1
 #endif
 #if VANILLA_AO_DARK == -1
 #undef VANILLA_AO_DARK
-										#define VANILLA_AO_DARK 0.8
+										#define VANILLA_AO_DARK 1.0
 #endif
 #if SIDE_SHADING_BRIGHT == -1
 #undef SIDE_SHADING_BRIGHT
@@ -837,7 +837,7 @@
 #endif
 #if OVERWORLD_BRIGHTNESS == -1
 #undef OVERWORLD_BRIGHTNESS
-										#define OVERWORLD_BRIGHTNESS 0.97
+										#define OVERWORLD_BRIGHTNESS 0.98
 #endif
 #if NETHER_BRIGHTNESS == -1
 #undef NETHER_BRIGHTNESS
@@ -1093,7 +1093,7 @@
 #endif
 #if WATER_TRANSPARENCY_DEEP == -1
 #undef WATER_TRANSPARENCY_DEEP
-										#define WATER_TRANSPARENCY_DEEP 0.15
+										#define WATER_TRANSPARENCY_DEEP 0.2
 #endif
 #if WATER_FOAM_ENABLED == -1
 #undef WATER_FOAM_ENABLED
@@ -1101,28 +1101,40 @@
 #endif
 #if WATER_FOAM_AMOUNT == -1
 #undef WATER_FOAM_AMOUNT
-										#define WATER_FOAM_AMOUNT 0.5
+										#define WATER_FOAM_AMOUNT 0.45
 #endif 
 
 
 
 
 
-#if WATER_COLOR_RED == -1
-#undef WATER_COLOR_RED
-										#define WATER_COLOR_RED 0.0
+#if WATER_SATURATION == -1
+#undef WATER_SATURATION
+										#define WATER_SATURATION 0.8
 #endif
-#if WATER_COLOR_GREEN == -1
-#undef WATER_COLOR_GREEN
-										#define WATER_COLOR_GREEN 0.2
+#if WATER_TINT_SHALLOW_RED == -1
+#undef WATER_TINT_SHALLOW_RED
+										#define WATER_TINT_SHALLOW_RED 0.8
 #endif
-#if WATER_COLOR_BLUE == -1
-#undef WATER_COLOR_BLUE
-										#define WATER_COLOR_BLUE 0.5
+#if WATER_TINT_SHALLOW_GREEN == -1
+#undef WATER_TINT_SHALLOW_GREEN
+										#define WATER_TINT_SHALLOW_GREEN 1.2
 #endif
-#if WATER_COLOR_AMOUNT == -1
-#undef WATER_COLOR_AMOUNT
-										#define WATER_COLOR_AMOUNT 0.2
+#if WATER_TINT_SHALLOW_BLUE == -1
+#undef WATER_TINT_SHALLOW_BLUE
+										#define WATER_TINT_SHALLOW_BLUE 1.1
+#endif
+#if WATER_TINT_DEEP_RED == -1
+#undef WATER_TINT_DEEP_RED
+										#define WATER_TINT_DEEP_RED 0.9
+#endif
+#if WATER_TINT_DEEP_GREEN == -1
+#undef WATER_TINT_DEEP_GREEN
+										#define WATER_TINT_DEEP_GREEN 0.9
+#endif
+#if WATER_TINT_DEEP_BLUE == -1
+#undef WATER_TINT_DEEP_BLUE
+										#define WATER_TINT_DEEP_BLUE 0.8
 #endif